+++ src/Exception/CrosspostConfigureException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace InstagramAPI\Exception;
+
+class CrosspostConfigureException extends RequestException
+{
+}